50 fix search bar

+++ apps/api/src/app/product/product.sql
@@ -2,7 +2,7 @@ CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name TEXT,
     description TEXT,
-    price INTEGER,
+    price DECIMAL,
     category TEXT[],
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP NOT NULL DEFAULT NOW()
